+++ src/docs_example/mod.rs
@@ -1,0 +1,3 @@
+pub mod app;
+pub mod errors;
+pub mod tui;
